refactor: add access controllers to services

+++ app/core/domain/services/file.py
@@ -4,17 +4,17 @@ from typing import Optional
 
 from app.core.domain.dto.common import Pagination
 from app.core.domain.dto.file import CreateFileDTO, ReloadFileDTO, FilesFindDTO
-from app.core.domain.exceptions.category import CategoryAccessDenied
 from app.core.domain.exceptions.file import FileNotFound, FileAccessDenied
 from app.core.domain.models.category import CategoryId
 from app.core.domain.models.file import File, FileId, RemoteFileId
 from app.core.domain.models.user import UserId
 from app.core.domain.services.internal import convert_to_filter_fields
+from app.core.interfaces.access import AccessController, Operation
 from app.core.interfaces.repository.file import (
     FileRepoSaver, FileRepoGetter, FileRepoFinder, FileRepoUpdater, FileRepoDeleter
 )
-from app.core.interfaces.usecase.category import CategoryGetter
 from app.core.interfaces.usecase import FileUsecase
+from app.core.interfaces.usecase.category import CategoryGetter
 
 UNDEFINED_FILE_ID = FileId(0)
 
@@ -45,7 +45,8 @@ class FileService(FileUsecase):
         self._deleter = deleter
         self._category_getter = category_getter
 
-    async def save_file(self, dto: CreateFileDTO) -> File:
+    async def save_file(self, dto: CreateFileDTO, access: AccessController) -> File:
+        access.ensure_have_access(Operation.file_create)
         file = File(
             UNDEFINED_FILE_ID,
             dto.title,
@@ -57,53 +58,80 @@ class FileService(FileUsecase):
         )
 
         if dto.category_id is not None:
-            category = await self._category_getter.get_category(CategoryId(dto.category_id), file.user_id)
+            category = await self._category_getter.get_category(
+                CategoryId(dto.category_id), access
+            )
+            access.ensure_own_category(category)
             file.category = category
 
         file.id = await self._saver.save_file(file)
         return file
 
-    async def get_file(self, file_id: FileId, user_id: Optional[UserId] = None) -> File:
+    async def get_file(self, file_id: FileId, access: AccessController) -> File:
         file = await self._getter.get_file(file_id)
         if file is None:
             raise FileNotFound(file_id)
-        _ensure_owner(file, user_id)
+        access.ensure_own_file(file)
 
         return file
 
-    async def set_category(self, file_id: FileId, category_id: CategoryId, user_id: UserId) -> File:
-        file = await self.get_file(file_id, user_id)
-        _ensure_owner(file, user_id)
+    async def set_category(
+        self,
+        file_id: FileId,
+        category_id: CategoryId,
+        access: AccessController
+    ) -> File:
+        access.ensure_have_access(Operation.file_edit)
+        file = await self.get_file(file_id, access)
 
-        category = await self._category_getter.get_category(category_id, user_id)
+        category = await self._category_getter.get_category(category_id, access)
+        access.ensure_own_category(category)
+
         file.category = category
         await self._updater.update_file(file)
         return file
 
-    async def update_title(self, file_id: FileId, title: str, user_id: UserId) -> File:
-        file = await self.get_file(file_id, user_id)
-        _ensure_owner(file, user_id)
+    async def update_title(
+        self,
+        file_id: FileId,
+        title: str,
+        access: AccessController
+    ) -> File:
+        access.ensure_have_access(Operation.file_edit)
+
+        file = await self.get_file(file_id, access)
+        access.ensure_own_file(file)
 
         file.title = title
         await self._updater.update_file(file)
         return file
 
-    async def reload_file(self, file_id: FileId, dto: ReloadFileDTO, user_id: UserId) -> File:
-        file = await self.get_file(file_id, user_id)
-        _ensure_owner(file, user_id)
+    async def reload_file(
+        self,
+        file_id: FileId,
+        dto: ReloadFileDTO,
+        access: AccessController
+    ) -> File:
+        access.ensure_have_access(Operation.file_edit)
+
+        file = await self.get_file(file_id, access)
+        access.ensure_own_file(file)
 
         file.remote_file_id = dto.remote_id
         file.remote_unique_id = dto.remote_unique_id
         file.type = dto.file_type
+
         if dto.title is not None:
             file.title = dto.title
 
         await self._updater.update_file(file)
         return file
 
-    async def delete_file(self, file_id: FileId, user_id: UserId):
-        file = await self.get_file(file_id, user_id)
-        _ensure_owner(file, user_id)
+    async def delete_file(self, file_id: FileId, access: AccessController):
+        access.ensure_have_access(Operation.file_delete)
+
+        file = await self.get_file(file_id, access)
+        access.ensure_own_file(file)
 
         await self._deleter.delete_file(file_id)
